+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
   - [#3656](https://github.com/bpftrace/bpftrace/pull/3656)
 - `len` now also accepts `ustack` and `kstack` as arguments
   - [#3769](https://github.com/bpftrace/bpftrace/pull/3769)
+- `blazesym` can now be used for kernel address symbolication, if configured & built with `USE_BLAZESYM`
+  - [#3760](https://github.com/bpftrace/bpftrace/pull/3760)
 #### Changed
 - `probe` builtin is now represented as a string type
   - [#3638](https://github.com/bpftrace/bpftrace/pull/3638)